+++ webui/src/store/effector/cluster.js
@@ -1,25 +1,10 @@
 // @flow
 import { createEvent } from 'effector'
 import { type ClusterPageState } from 'src/store/reducers/clusterPage.reducer';
-
-export type ClusterDisableServersSuggestion = {
-  uuid: string
-}
-
-export type ClusterRefineURISuggestion = {
-  uri_new: string,
-  uri_old: string,
-  uuid: string
-};
-
-export type ClusterSuggestions = {
-  refine_uri?: ClusterRefineURISuggestion[],
-  disable_servers?: ClusterDisableServersSuggestion[]
-};
-
+import type { Suggestions } from 'src/generated/graphql-typing';
 export type ClusterState = {
   ...$Exact<ClusterPageState>,
-  suggestions?: ClusterSuggestions
+  suggestions?: Suggestions
 };
 
 export const clusterPageMount = createEvent<mixed>('cluster page mount');